requirements.txt
google-adk
requests
streamlit
scapy
reportlab
matplotlib
pandas
pyyaml
tqdm